requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0
motor>=3.3.0
pymongo>=4.6.0
flask>=3.0.0
